2010-04-29	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libnetworking/libc/map_v4v6.c: Use uintptr_t instead of u_long for
	better 16bit target compliance.

+++ cpukit/libnetworking/libc/map_v4v6.c
@@ -108,7 +108,7 @@ _map_v4v6_hostent(
 	hp->h_addrtype = AF_INET6;
 	hp->h_length = IN6ADDRSZ;
 	for (ap = hp->h_addr_list; *ap; ap++) {
-		int i = sizeof(align) - ((u_long)*bpp % sizeof(align));
+		int i = sizeof(align) - ((uintptr_t)*bpp % sizeof(align));
 
 		if (*lenp < (i + IN6ADDRSZ)) {
 			/* Out of memory.  Truncate address list here.  XXX */
